+++ corbeans-spring-boot-corda/src/main/kotlin/com/github/manosbatsis/corbeans/spring/boot/corda/rpc/EagerNodeRpcConnection.kt
@@ -30,7 +30,7 @@ import org.slf4j.LoggerFactory
  * @property proxy The RPC proxy.
  */
 open class EagerNodeRpcConnection(
-        nodeParams: NodeParams): NodeRpcConnection(nodeParams) {
+        nodeParams: NodeParams): AbstractNodeRpcConnection(nodeParams) {
 
     companion object {
         private val logger = LoggerFactory.getLogger(EagerNodeRpcConnection::class.java)